requirements.txt
Flask==1.0.2
pymongo==3.7.1
flask_restful==0.3.8
python_dateutil==2.8.1
